requirements.txt
python-dotenv>=1.2.1
feedparser>=6.0.12
psycopg2-binary>=2.9.11
Jinja2>=3.1.6
requests>=2.32.5
Flask>=3.1.2